Do not use Optional for fields
https://stackoverflow.com/a/26328555/3925216

Of course, people will do what they want. But we did have a clear intention when adding this feature,
and it was not to be a general purpose Maybe type, as much as many people would have liked us to do
so. Our intention was to provide a limited mechanism for library method return types where there
needed to be a clear way to represent "no result", and using null for such was overwhelmingly likely
to cause errors.